change sound config
